+++ spec/active_record/trino/config_spec.rb
@@ -91,6 +91,21 @@ RSpec.describe ActiveRecord::Trino::Config do
     end
   end
 
+  describe ".bulk_column_reflection?" do
+    it "defaults to false" do
+      expect(described_class.bulk_column_reflection?(base_config)).to be false
+    end
+
+    it "is true when bulk_column_reflection: true" do
+      expect(described_class.bulk_column_reflection?(base_config.merge(bulk_column_reflection: true))).to be true
+    end
+
+    it "accepts string keys (as database.yml usually provides)" do
+      string_keyed = base_config.merge(bulk_column_reflection: true).transform_keys(&:to_s)
+      expect(described_class.bulk_column_reflection?(string_keyed)).to be true
+    end
+  end
+
   describe ".slow_query_threshold" do
     it "defaults to 20.0 seconds" do
       expect(described_class.slow_query_threshold(base_config)).to eq(20.0)